Add support for certificate chains
This patch adds the ability to both sign and verify
signatures that contain a certificate chain. The root
of the chain is expected to be provided separately via
an x509.CertPool passed to the Verifier.

The added FirefoxAddon test case uses a certificate
chain to check the addon signature against the Firefox
Root CA.