Fix encryption support for AES-128-CBC
Previously the encrypt method was grouping together multiple algorithms
and handling them the same. In the decryption code this somewhat
makes sense because the crypto.Cipher will take care of figuring out
which algorithm it needs to use. When encrypting though if AES-128-CBC
is requested we were encrypting as AES-128-GCM which didn't make sense.